feat: make Bitcoin indexer use timeout from config
Previously, this timeout was hard-coded to 300s, and when a problem
happens, it takes way too long to timeout and try again.

+++ stackslib/src/config/mod.rs
@@ -1350,7 +1350,7 @@ pub struct BurnchainConfig {
     /// ---
     /// @default: `60`
     /// @units: seconds
-    pub timeout: u32,
+    pub timeout: u64,
     /// The network "magic bytes" used to identify packets for the specific bitcoin
     /// network instance (e.g., mainnet, testnet, regtest). Must match the magic
     /// bytes of the connected bitcoin node.
@@ -1752,7 +1752,7 @@ pub struct BurnchainConfigFile {
     pub username: Option<String>,
     pub password: Option<String>,
     /// Timeout, in seconds, for communication with bitcoind
-    pub timeout: Option<u32>,
+    pub timeout: Option<u64>,
     pub magic_bytes: Option<String>,
     pub local_mining_public_key: Option<String>,
     pub process_exit_at_block_height: Option<u64>,
